move 4 files, create 1 file and delete 3 files
