--- 023_build_an_application/testutils/http_helpers.go
@@ -1,32 +1,0 @@
-// Package testutils provides utility functions for testing for HTTP handlers
-// and responses, temporary files, and assertions in tests.
-package testutils
-
-import (
-	"net/http/httptest"
-	"testing"
-)
-
-// AssertResponseBody checks if the actual response body matches the expected value.
-func AssertResponseBody(t testing.TB, got, want string) {
-	t.Helper()
-	if got != want {
-		t.Errorf("response body is wrong, got %q want %q", got, want)
-	}
-}
-
-// AssertStatus checks whether the actual HTTP status code matches the expected one.
-func AssertStatus(t testing.TB, got, want int) {
-	t.Helper()
-	if got != want {
-		t.Errorf("did not get correct status, got %d, want %d", got, want)
-	}
-}
-
-// AssertContentType checks that the Content-Type header matches the expected value.
-func AssertContentType(t testing.TB, response *httptest.ResponseRecorder, want string) {
-	t.Helper()
-	if response.Result().Header.Get("content-type") != want {
-		t.Errorf("response did not have content-type of %s, got %v", want, response.Result().Header)
-	}
-}
